removed old client + fixed all warnings
