+++ relays/client-substrate/src/rpc.rs
@@ -16,15 +16,9 @@
 
 //! The most generic Substrate node RPC interface.
 
-use async_trait::async_trait;
+use crate::{Chain, TransactionStatusOf};
 
-use crate::{Chain, ChainWithGrandpa, TransactionStatusOf};
-
-use jsonrpsee::{
-	core::{client::Subscription, RpcResult},
-	proc_macros::rpc,
-	ws_client::WsClient,
-};
+use jsonrpsee::{core::RpcResult, proc_macros::rpc};
 use pallet_transaction_payment_rpc_runtime_api::FeeDetails;
 use sc_rpc_api::{state::ReadProof, system::Health};
 use sp_core::{
@@ -106,13 +100,6 @@ pub(crate) trait SubstrateState<C> {
 	) -> RpcResult<ReadProof<C::Hash>>;
 }
 
-/// RPC methods that we are using for a certain finality gadget.
-#[async_trait]
-pub trait SubstrateFinalityClient<C: Chain> {
-	/// Subscribe to finality justifications.
-	async fn subscribe_justifications(client: &WsClient) -> RpcResult<Subscription<Bytes>>;
-}
-
 /// RPC methods of Substrate `grandpa` namespace, that we are using.
 // TODO: replaced `ChainWithGrandpa` with `Chain`. Is it possible to revert?
 #[rpc(client, client_bounds(C: Chain), namespace = "grandpa")]
@@ -122,16 +109,6 @@ pub(crate) trait SubstrateGrandpa<C> {
 	async fn subscribe_justifications(&self);
 }
 
-/// RPC finality methods of Substrate `grandpa` namespace, that we are using.
-pub struct SubstrateGrandpaFinalityClient;
-// TODO: replaced `ChainWithGrandpa` with `Chain`. Is it possible to revert?
-#[async_trait]
-impl<C: Chain> SubstrateFinalityClient<C> for SubstrateGrandpaFinalityClient {
-	async fn subscribe_justifications(client: &WsClient) -> RpcResult<Subscription<Bytes>> {
-		SubstrateGrandpaClient::<C>::subscribe_justifications(client).await
-	}
-}
-
 // TODO: Use `ChainWithBeefy` instead of `Chain` after #1606 is merged
 /// RPC methods of Substrate `beefy` namespace, that we are using.
 #[rpc(client, client_bounds(C: Chain), namespace = "beefy")]
@@ -139,16 +116,6 @@ pub(crate) trait SubstrateBeefy<C> {
 	/// Subscribe to BEEFY justifications.
 	#[subscription(name = "subscribeJustifications", unsubscribe = "unsubscribeJustifications", item = Bytes)]
 	async fn subscribe_justifications(&self);
-}
-
-/// RPC finality methods of Substrate `beefy` namespace, that we are using.
-pub struct SubstrateBeefyFinalityClient;
-// TODO: Use `ChainWithBeefy` instead of `Chain` after #1606 is merged
-#[async_trait]
-impl<C: Chain> SubstrateFinalityClient<C> for SubstrateBeefyFinalityClient {
-	async fn subscribe_justifications(client: &WsClient) -> RpcResult<Subscription<Bytes>> {
-		SubstrateBeefyClient::<C>::subscribe_justifications(client).await
-	}
 }
 
 /// RPC methods of Substrate `system` frame pallet, that we are using.
